Multi language support for installer.
Installer will now detect the existence of language files in BB_LANG_DIR and return the option on step zero to select a language to use during installation.

A language can also be selected for the installation independently from the installer language. So a person can install a foreign forum in their chosen language. A true "border" case. Boom-tish.

Also contained in this commit is the long awaited RTL installation CSS.

Fixes #839

git-svn-id: https://bbpress.svn.wordpress.org/trunk@1370 9866e705-20ec-0310-96e7-cbb4277adcfb