Change 'bbp-twentyten' theme from child to parent and include all of the standard twentyten theme files. Also wrap function_exists() checks around all functions in bbp-twentyten to allow them to be overloaded.
Remove the 'bbp_' prefix form/topic/reply post types and include a routine in the database updater to handle this. The new post_type's are 'forum', 'topic', and 'reply' respectively.

Rename 'bbp_topic_tag' taxonomy ID to 'topic-tag' and include a routine in the database updater to handle this.

Change underscores in theme files to hyphens and remove 'bbp' prefix where applicable to comply with WordPress file standards.

Update possible template file location options in bbp-general-functions.php for all standard template files. Include option for 'bbpress' or 'forums' subfolders inside the theme directory. This takes advantage of a known anomaly in the get_template_part() function that allows a subfolder to be passed.

git-svn-id: https://bbpress.svn.wordpress.org/branches/plugin@2943 9866e705-20ec-0310-96e7-cbb4277adcfb