Improved trash support for new topics/replies.
If the parent of a new topic/reply is in the trash, automatically trash the new topic/reply. In the case of new trash replies, add that reply to the _bbp_pre_trashed_replies array, so that it is restored when the topic is restored. This bug comes about because capable users can reply to trashed topics.

Also handle view=all redirect when posting new topics/replies.

git-svn-id: https://bbpress.svn.wordpress.org/branches/plugin@3240 9866e705-20ec-0310-96e7-cbb4277adcfb