bbPress 1.x: Allow log-in attempt even if we think the user is already logged in.
If your session cookie's internal expiration time has passed, we give you a grace period of one hour for POST and XHR requests. If within that hour you're trying to log in again, the grace period goes into effect for the POST. Then bb-login.php kicks you out because it thinks you're already logged in, instead of setting a new cookie for you.

So long, heisenbug.


git-svn-id: https://bbpress.svn.wordpress.org/branches/1.1@5305 9866e705-20ec-0310-96e7-cbb4277adcfb