bbPress 1.x: Call bb_get_user() from bb_get_user_by_nicename().
This ensures that this gem runs from bb_get_user() on profile pages: "Re calculate the user's meta in case we're pulling from a value cached on another site". So far, we know this has manifested itself as causing incorrect pagination for activity on profile pages, based on which site did the initial caching.

In an ideal world, these "short" keys would use a magic getter on BP_User, but user objects in bbPress 1.x are mostly stdClass objects in practice. Fun times.


git-svn-id: https://bbpress.svn.wordpress.org/branches/1.1@5595 9866e705-20ec-0310-96e7-cbb4277adcfb