Users: Remove `is_email()` check from user query parsing.
This commit removes the ability to load a user profile via their email address. This feature (added in r2917) was a convenience mechanism to make finding a user within the forums with a known address easier, but times have changed and we now consider this to be a privacy concern by default.

Note: this will break backwards-compatibility for installations that may have used this unpublicized (and regrettably unwise) feature.

Fixes #2828. Hat-tip netweb. For 2.5.8 (2.5 branch)

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.5@5814 9866e705-20ec-0310-96e7-cbb4277adcfb