Core: Support for `WP_Hook` in `bbp_get_template_stack()`
The proposed `WP_Hook` #WP17817 breaks backward compatibility for plugins that
attempt to manipulate `$wp_filter` in certain ways. For bbPress, it means
that we cannot directly modify the callbacks that are currently registered as
an array at `$wp_filter[ $tag ]`, in `bbp_get_template_stack()`. Instead, when
`WP_Hook` is not available, we assign the callbacks to a variable (leveraging
`WP_Hook`'s `ArrayAccess`) and manipulate the variable.

Props jbrinley. Fixes #2871


git-svn-id: https://bbpress.svn.wordpress.org/trunk@5944 9866e705-20ec-0310-96e7-cbb4277adcfb