Tests: Rely on WP Core's version of `go_to()`.
Our version of `go_to()`, inherited from BuddyPress was very messy, and since it was copied a few years ago, it has been refactored in WordPress. We can remove this duplication which makes for a cleaner method, and rely on the upstream version.

This also helps avoid false positives in tools that check for use of deprecated WordPress functions.

There was a conditional block in the old go_to() that was for compatibility with WordPress <3.9.

See also: buddypress:changeset:11254

Fixes #3015.
Props DJPaul, netweb.


git-svn-id: https://bbpress.svn.wordpress.org/trunk@6231 9866e705-20ec-0310-96e7-cbb4277adcfb