Moderators: Performance optimization in `bbp_is_user_forum_moderator()`.
Rather than query for all forums a user can moderate, query for the users that can moderate the forum. This data is likely already hot in the meta-data cache from forums being loaded previously, and eliminates the need to query for forum IDs that are unrelated to the one being checked.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6406 9866e705-20ec-0310-96e7-cbb4277adcfb