Engagements: Tracking is not just for public or closed topics.
This commit updates the engagement upgrade tool and recalculation helper functions to not be limited to public & closed post statuses. Engagements can happen on private or hidden topics also, including trash, spam, and so on.

The non-visible topics that a user is engaged in should be filtered out as needed when queried based on `post_status` or other various meta-data.

See #3068.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6525 9866e705-20ec-0310-96e7-cbb4277adcfb