Super Moderators: prefer `edit_user` with an ID over `edit_users` without one.
This change makes sure we know exactly which user ID the user is trying to edit. In cases where there is no ID available, we'll try to guess at the displayed user ID, and if that's empty too, we'll do nothing.

This allows the "User Role" template part to load properly for Super Moderators.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6575 9866e705-20ec-0310-96e7-cbb4277adcfb