Cache: make `bbp_clean_post_cache()` an action-only function.
Instead of calling `bbp_clean_post_cache()` directly, we'll call `clean_post_cache()` as per normal and hook `bbp_clean_post_cache()` to the end of it. This reduces a bunch of superfluous cache invalidation from occurring back-to-back.

We'll also only ever update the `last_changed` cache key when we've reached the forum-root. This makes sure that subsequent recursive calls up the `post_parent` tree are dealing with the same `last_changed` value until the end.

This change will reduce the number of cache invalidation calls by a large amount, improving functional performance for database writes to the post & postmeta tables.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6585 9866e705-20ec-0310-96e7-cbb4277adcfb