Styling: No border on forums links.
An increasing number of themes are putting borders on links by default to simulate an underline, which causes every link in the forums to be styled in unanticipated ways.

This changeset ensures that links in the entry-content section in forums do not have faux-underlines or excessive attention drawn to them.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6598 9866e705-20ec-0310-96e7-cbb4277adcfb