Users: Use `bbp_parse_args()` in user content loops.
This commit changes the function signatures of several wrappers for `bbp_has_` functions away from simply accepting a `$user_id` to accepting an array of arguments, in a fully backwards compatible way. It also updates the surrounding documentation to more accurately describe what is returned and why.

This allows the arguments used within these functions to be explicitly filtered, passively or aggressively overridden, or bypassed entirely.



git-svn-id: https://bbpress.svn.wordpress.org/trunk@6619 9866e705-20ec-0310-96e7-cbb4277adcfb