Tools: include users with "unexpected" roles in `bbp_admin_repair_user_roles()`
This change makes sure that if a user has an unexpectedly missing or predictably broken role entry, they are still mapped to the `$default_role` for the current site, and fixes a problem with the role repair tool where users with the most broken roles were still never actually repaired.

Trunk, for 2.6.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6668 9866e705-20ec-0310-96e7-cbb4277adcfb