Core: Add `source` parameter to API registrations.
This change makes it easier to target bbPress-specific post types & statuses, and taxonomies. `source` may not be the final key, but thankfully all of these APIs accept additional keys, and if support is added to core eventually, we can easily migrate over to it without problems.

Trunk, for 2.6. See: https://core.trac.wordpress.org/ticket/41739

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6672 9866e705-20ec-0310-96e7-cbb4277adcfb