Users: first pass at forum role selection on invite/add-new.
This change adds the ability to set a specific forum role when adding a new user. If no valid role is found or saved, the default role will be used by default. There is likely more to do here, to ensure auto-role continues to work as intended as well.

Trunk, for 2.6. See: #3157.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6675 9866e705-20ec-0310-96e7-cbb4277adcfb