Template: rework post class assignments.
This change avoids multiple reassignments to the same `$classes` variable name, and instead tries to name variables logically and merge them together when necessary. The performance difference is nil, as 'array_merge()` will perform similarly to how each array was reshaped when new classes would be added, but the human difference is only positive, from having clearer and easier to understand logic.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6683 9866e705-20ec-0310-96e7-cbb4277adcfb