Intercept: first pass intercept API.
This change introduces 3 new functions for generating a default intercept value and comparing against it in specific places. If the return value differs from the default intercept value, we know that function call was intercepted by a filter, and that value will become the new return value without executing the remaining part of the function.

See #3184.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6751 9866e705-20ec-0310-96e7-cbb4277adcfb