Ajax: send a proper `400` response when nothing happens.
This change adds nocache and noindex headers to theme-side AJAX handling, adds support for pre-sanitized `$action` parameters, and checks that the action-hook actually has actions hooked to it before attempting to execute it.

Similar to `admin-ajax.php` a `400` response is now returned when an AJAX request is attempted that does not have a handler.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6755 9866e705-20ec-0310-96e7-cbb4277adcfb