Converter: remove `WP_SETUP_CONFIG` constant setting from converter process.
This introduces a helper database class to avoid directly connecting to the external database. Instead, we'll attempt to control that connection, and provide feedback to the user in the event a connection cannot be made. The `WP_SETUP_CONFIG` was causing calls to the options API to fail, resulting in broken calculations and invalid offsets.

This commit includes some general load order clean-up, which also fixes a regression causing fatal errors when attempting to upgrade converted user passwords from the old platform. It also fixes the condition where boundaries being converted would display beyond their maximums.

Fixes #3191.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6785 9866e705-20ec-0310-96e7-cbb4277adcfb