Tools: reduce the number of calls to `bbp_is_large_install()`.
This commit refactors the update code to only call `bbp_is_large_install()` once, a function which ironically enough is costly to call on very large installations, so calling it 3 times defeats the purpose, but calling it one time is the minimum requirement to defer upgrades & repairs for keymaster interaction.

See #3052.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6935 9866e705-20ec-0310-96e7-cbb4277adcfb