Replies: ensure visual editor works correctly with hierarchical replies.
This commit updates reply.js in the default theme to include a bit of element juggling that prevents TinyMCE from freezing up when it is relocated on the page.

Before this commit, clicking "Reply" while having the Visual Editor enabled would cause it to not be able to be typed inside of. Now, when clicking commit, TinyMCE is correctly reinvoked in its new location, and the page will expediently scroll to that new position, fixing a related UX bug where it was possible for the reply form to get lost inside of deep hierarchies.

Props chherbst. Fixes #2646. For 2.6.2, branches/2.6.



git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@6965 9866e705-20ec-0310-96e7-cbb4277adcfb