Slugs: add support for configuring the "Edit" slug.
This commit adds supporting functions and UI for allowing the `edit` slug to be modified. This omission was originally intentional, as the rewrite rules and pretty permalink support were not fully implemented. Now that they are in 2.6, there is no reason not to allow this slug to be modified.

Related changes include swapping out the new slug setting for the rewrite ID in a few places, and removing a single hardcoded reference to `'edit'` that was also overlooked.

Props espellcaste for noticing. Fixes #3285. For 2.6.2, branches/2.6.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@6966 9866e705-20ec-0310-96e7-cbb4277adcfb