Default Styling: Optionally support an RTL variant of bbpress.css.
This commit allows for either/or `bbpress.css` or `bbpress-rtl.css` to exist and be enqueued as styling in the template stack, fixing a regression from 2.6.0 where the RTL CSS was expected to be appended to the end of the file.

This allows for both files to be included if they both exist, or one of them if only one does, effectively covering all bases for 2.5 and 2.6 support.

Props scmsteve for persistence. Fixes #3218. For 2.6.2, branches/2.6.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@6969 9866e705-20ec-0310-96e7-cbb4277adcfb