Capabilities: break instead of return;
This commit ensures that the relative `_meta_caps` filters continue to be executed within their expected contexts.

Previous to this change, certain mapped capability checks would bail out early and return `$caps` without the ability to filter the results.

With this change, mapped capabilities can now be filtered more logically and inline with the rest of the surrounding logic.

For 2.6.2.



git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@6972 9866e705-20ec-0310-96e7-cbb4277adcfb