BuddyPress: explicitly validate IDs when editing Group forum topics & replies.
This commit adds methods to validate that the forum IDs and reply-to IDs for topics & replies are within the accepted ranges for the specific Group Forum they are being edited inside of.

In addition, the `moderate_forum` mapped meta capability is removed, and the broader `moderate` capability will continue to cover its use case. This capability was not intended to be used directly, and doing so incorrectly would trigger unintended and infinite recursion.

For 2.7, trunk

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7060 9866e705-20ec-0310-96e7-cbb4277adcfb