Theme Compatibility: Avoid compatibility issues between MySQL versions.
This commit introduces a new function - `bbp_get_empty_datetime()` - which returns a valid, filterable, "empty" MySQL formatted datetime value.

This new function is used inside the theme compatibility API to ensure that global `WP_Post` datetime values are not "0" to better support MySQL 8 and higher, which is dropping support for `NO_ZERO_DATES`.

Props GDragoN.

In trunk, for 2.7.0.

See #3354.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7099 9866e705-20ec-0310-96e7-cbb4277adcfb