BuddyPress: Mark all replies when marking topic notifications as read.
This commit fixes a regression - introduced in r6845 - that was causing marking topic notifications as read to fail. It fixes it by looping through all replies to a topic and attempting to mark them all individually. It is not a particularly optimized approach, but it does resolve the regression in such a way that accounts for both topic IDs and reply IDs.

In trunk, for 2.7.0.

Fixes #3378.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7108 9866e705-20ec-0310-96e7-cbb4277adcfb