Akismet: fix spam check not sending author info for logged in users
This commit uses `bbp_has_errors()` to catch whether anonymous information exists or not, and falls back to the currently logged in user otherwise (anonymous has priority due to moderator ability to edit topics & replies).

This commit also improves the readability of a few lengthy function calls, and adds empty() checks to all of the related array key touches.

In trunk, for 2.7.

Props procifer.

See #3368.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7129 9866e705-20ec-0310-96e7-cbb4277adcfb