Replies: prevent adding empty values to pre-spam & pre-trash meta data.
This commit filters empty values from meta data before saving them, as well as removing a simple type-cast and replacing it with a more intentional is_array() check. This new approach works because get_post_meta() will call maybe_unserialize() and always return an array if it exists, or return an empty string if it does not. If it's not an array, we know it should be an empty one.

In branches/2.6, for 2.6.6. See #3409.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7140 9866e705-20ec-0310-96e7-cbb4277adcfb