Replies: Allow moderators to reply to pending topics.
This commit adds a check inside of `bbp_current_user_can_access_create_reply_form()` for the `edit_topic` capability, and if the user is capable will now allow them to access that form.

This will come in particularly handy for forum moderators who want to reply to topics before they are approved, or to provide a reason to the original poster why a topic will not be approved but without making it public.

In trunk, for 2.7. Fixes #3309.

Props dd32, man4toman.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7158 9866e705-20ec-0310-96e7-cbb4277adcfb