Post Statuses: Audit forum/topic/reply, new/edit handlers.
This change normalizes the approaches between post types and actions, to ensure predictable behavior and avoid PHP notices from undefined status variables in certain situations.

It also adds a capability check to the `bbp_topic_status` condition, so that it is not engaged when the UI was not exposed in the posted form.

Props dd32.

In trunk for 2.7.0. See #3420.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7191 9866e705-20ec-0310-96e7-cbb4277adcfb