Akismet: Remove "publish" status restriction in check_post().
This commit ensures that posts of any status are sent to Akismet, even if they had previously been published and are now being edited. The intent of this change is to remove a restriction without changing the behavior from a user's perspective.

Also fixes a typo and an incorrectly typed return value.

Props procifer, douglsmith.

In trunk for 2.7.0. See #3410.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7193 9866e705-20ec-0310-96e7-cbb4277adcfb