Content: add `post_parent` to forum & topic duplicate checks.
This change limits the scope of "duplicate content" to its parent context, meaning that replies must be to the same topic, topics must be in the same forum, and forums must be in the same sub-forum (or root) to be considered a duplicate.

This better implements the spirit of this feature which is to prevent accidental/unintended duplicates, and is not intended to simply prevent globally similar content by the same author.

Props r-a-y, johnjamesjacoby.

In branches/2.6, for 2.6.7.

Fixes #3418.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7208 9866e705-20ec-0310-96e7-cbb4277adcfb