Users: add some error checking to `get_userdata()` calls in profile link functions.
This change ensures that when attempting to link to user profiles when the `display_name` is the inner HTML of the link, that the "Anonymous" alternate is used whenever the user cannot be found.

It also bails early – returning `false` – if the user account simply does not exist, consistent with upstream calls to `bbp_get_user_id()`. (Functions calling these functions have a responsibility to appropriately handle boolean return values.)

In branches/2.6, for 2.6.7.

See #3440.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7213 9866e705-20ec-0310-96e7-cbb4277adcfb