Appearance: only add `no-js` body class on bbPress pages.
This change addresses a regression introduced in r7190, causing the `bbpress` body class to unintentionally be added to all pages triggering layout issues in some WordPress themes.

This commit makes sure that the `bbpress` body class is only added when intended, and moves the `no-js` body class into the same conditional check.

In trunk for 2.7. See #3443.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7220 9866e705-20ec-0310-96e7-cbb4277adcfb