Appearance: only add `no-js` body class on bbPress pages.
This change addresses a regression introduced in r7190, causing the `bbpress` body class to unintentionally be added to all pages triggering layout issues in some WordPress themes.

This commit makes sure that the `bbpress` body class is only added when intended, and moves the `no-js` body class into the same conditional check.

In branches/2.6 for 2.6.8. Fixes #3443.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7221 9866e705-20ec-0310-96e7-cbb4277adcfb