Theme Compat: improve "no-js" body class swapping.
This change introduces a new helper function (`bbp_swap_no_js_body_class()`) which is used to output the `<script>` tag that was previously inside of the single template part that needed it.

It also adds a prefix (`bbp`) to those "no-js" body classes, with the intention of avoiding conflicts with other plugins & themes that may also be using the `no-js` body class for their own purposes.

Props dd32, r-a-y.

In branches/2.6 for 2.6.10. Fixes #3445.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7230 9866e705-20ec-0310-96e7-cbb4277adcfb