Moderation: use the correct `post_status` when untrashing a topic/reply.
This change fixes a bug/regression (since WordPress 5.6.0) that was causing untrashed topics & replies to use an unintended `post_status` value.

It fixes it by adding a new function ('bbp_fix_untrash_post_status()') and hooking it to the `wp_untrash_post_status` filter in WordPress, and overriding the results as needed.

Props r-a-y.

In branches/2.6 for 2.6.10. Fixes #3433.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7234 9866e705-20ec-0310-96e7-cbb4277adcfb