Moderation: Make moderated or disallowed key check case-insensitive for non-Latin words.
The `bbp_check_for_moderation()` function expects to be case-insensitive, but that only works for words using Latin script and consisting of ASCII characters.

This commit adds the Unicode flag to the regular expression used for the check in these functions, so that both pattern and subject can be treated as UTF-8 strings.

See: https://www.php.net/manual/en/reference.pcre.pattern.modifiers.php

In branches/2.6, for 2.6.14.

Fixes #3640.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7304 9866e705-20ec-0310-96e7-cbb4277adcfb