Extend - BuddyPress: Improve support for BuddyPress 12.0 and higher.
This change fixes fatal errors when BuddyPress Group/Member functions (deprecated in 12.0) are not included.

(Note that this commit differs from the proposed patch in that it does not modify the action-order of existing integrations, hence it does not rely on BuddyPress versions greater than 12.0 being in-use, and is safe for use with older BuddyPress versions.)

Props boonebgorges, imath, robin-w.

In trunk, for 2.7.

Fixes #3576.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7316 9866e705-20ec-0310-96e7-cbb4277adcfb