Extend - Akismet: avoid timeouts in the `delete_orphaned_spam_meta()` method.
This change modifies the SQL used to query for Akismet post-meta keys, such that it will only retrieve rows where `meta_key` is like `akismet_`.

This results in a less-optimized database query, but circumvents a bug where unnecessary rows were being looped through in a way that would never finish.

Props terresquall.

In trunk, for 2.7.

Fixes #3580.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7325 9866e705-20ec-0310-96e7-cbb4277adcfb