Theme Compat: remove reference to non-existent function.
This change removes a function call to `display_reply_index()` which does not exist, and replaces it with an empty string.

This condition will not fire under normal circumstances as the Reply post type does not support archives by default. The chances of this fatal error happening are near zero, but we can easily avoid it fully by returning an empty string instead.

Props sirlouen.

In branches/2.6, for 2.6.14.

Fixes #3616.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7328 9866e705-20ec-0310-96e7-cbb4277adcfb