Forums: swap out `bbp_is_user_keymaster()` check for `moderate` capability when checking the `edit_forums` and `edit_others_forums` mapped forum caps.
This commit fixes a bug that was causing the top-level admin-area "Forums" menu to not appear for logged in users who were assigned the "Moderator" role.

Moderators are intended to be like Editors, capable of moderating all forum-specific content. Sometimes that includes editing forum titles, descriptions, slugs, or creating a new forum and moving a bunch of existing topics into it.

Note: this change is largely a visual one, as Moderators have had the ability to edit forums since bbPress 2.2 when BuddyPress Group Forum support was added, if they already had the edit URL.

In trunk, for 2.7.

Fixes #3646.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7346 9866e705-20ec-0310-96e7-cbb4277adcfb