Core - Canonical: Convert `bbp_redirect_canonical()` into a sub-action.
This change abstracts the existing internal canonical redirect logic into new dedicated functions so that they can be individually improved or customized in the future.

Pagination and Editing theme-side require intercepting and preventing the core redirection when requesting certain URLs, and there are likely to be new conditions in the future.

In trunk, for 2.7.

Fixes #3648.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7348 9866e705-20ec-0310-96e7-cbb4277adcfb