Akismet: improve verbiage in history log items.
When Akismet catches spam for moderators, and moderators are able to bypass spam, a second log entry is inserted into the history to explain why, and the previous verbiage was not very clear.

This commit attempts to provide some clarity, while also respecting that this specific code is filterable, and that the spam recommendation could be overruled by third-party plugins for any reason.

This commit also adds some hooks by request, and improves some related code docs.

Props johnjamesjacoby, r-a-y, sirlouen.

In trunk, for 2.7.

See r7353. Fixes #3434.

git-svn-id: https://bbpress.svn.wordpress.org/trunk@7356 9866e705-20ec-0310-96e7-cbb4277adcfb